+++ subsys/bluetooth/controller/hci/hci.c
@@ -482,6 +482,86 @@ static void le_set_random_address(struct net_buf *buf, struct net_buf **evt)
 	ccst->status = 0x00;
 }
 
+static void le_read_wl_size(struct net_buf *buf, struct net_buf **evt)
+{
+	struct bt_hci_rp_le_read_wl_size *rp;
+
+	rp = cmd_complete(evt, sizeof(*rp));
+	rp->status = 0x00;
+
+	rp->wl_size = 8;
+}
+
+static void le_clear_wl(struct net_buf *buf, struct net_buf **evt)
+{
+	struct bt_hci_evt_cc_status *ccst;
+
+	ll_filter_clear();
+
+	ccst = cmd_complete(evt, sizeof(*ccst));
+	ccst->status = 0x00;
+}
+
+static void le_add_dev_to_wl(struct net_buf *buf, struct net_buf **evt)
+{
+	struct bt_hci_cp_le_add_dev_to_wl *cmd = (void *)buf->data;
+	struct bt_hci_evt_cc_status *ccst;
+	u32_t status;
+
+	status = ll_filter_add(cmd->addr.type, &cmd->addr.a.val[0]);
+
+	ccst = cmd_complete(evt, sizeof(*ccst));
+	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_MEM_CAPACITY_EXCEEDED;
+}
+
+static void le_rem_dev_from_wl(struct net_buf *buf, struct net_buf **evt)
+{
+	struct bt_hci_cp_le_rem_dev_from_wl *cmd = (void *)buf->data;
+	struct bt_hci_evt_cc_status *ccst;
+	u32_t status;
+
+	status = ll_filter_remove(cmd->addr.type, &cmd->addr.a.val[0]);
+
+	ccst = cmd_complete(evt, sizeof(*ccst));
+	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED;
+}
+
+static void le_encrypt(struct net_buf *buf, struct net_buf **evt)
+{
+	struct bt_hci_cp_le_encrypt *cmd = (void *)buf->data;
+	struct bt_hci_rp_le_encrypt *rp;
+	u8_t enc_data[16];
+
+	ecb_encrypt(cmd->key, cmd->plaintext, enc_data, NULL);
+
+	rp = cmd_complete(evt, sizeof(*rp));
+
+	rp->status = 0x00;
+	memcpy(rp->enc_data, enc_data, 16);
+}
+
+static void le_rand(struct net_buf *buf, struct net_buf **evt)
+{
+	struct bt_hci_rp_le_rand *rp;
+	u8_t count = sizeof(rp->rand);
+
+	rp = cmd_complete(evt, sizeof(*rp));
+	rp->status = 0x00;
+
+	bt_rand(rp->rand, count);
+}
+
+static void le_read_supp_states(struct net_buf *buf, struct net_buf **evt)
+{
+	struct bt_hci_rp_le_read_supp_states *rp;
+
+	rp = cmd_complete(evt, sizeof(*rp));
+	rp->status = 0x00;
+
+	sys_put_le64(0x000003ffffffffff, rp->le_states);
+}
+
+#if defined(CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV)
 static void le_set_adv_param(struct net_buf *buf, struct net_buf **evt)
 {
 	struct bt_hci_cp_le_set_adv_param *cmd = (void *)buf->data;
@@ -543,6 +623,7 @@ static void le_set_adv_enable(struct net_buf *buf, struct net_buf **evt)
 	ccst = cmd_complete(evt, sizeof(*ccst));
 	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED;
 }
+#endif /* CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV */
 
 static void le_set_scan_param(struct net_buf *buf, struct net_buf **evt)
 {
@@ -619,73 +700,6 @@ static void le_create_conn_cancel(struct net_buf *buf, struct net_buf **evt)
 	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED;
 }
 
-static void le_read_wl_size(struct net_buf *buf, struct net_buf **evt)
-{
-	struct bt_hci_rp_le_read_wl_size *rp;
-
-	rp = cmd_complete(evt, sizeof(*rp));
-	rp->status = 0x00;
-
-	rp->wl_size = 8;
-}
-
-static void le_clear_wl(struct net_buf *buf, struct net_buf **evt)
-{
-	struct bt_hci_evt_cc_status *ccst;
-
-	ll_filter_clear();
-
-	ccst = cmd_complete(evt, sizeof(*ccst));
-	ccst->status = 0x00;
-}
-
-static void le_add_dev_to_wl(struct net_buf *buf, struct net_buf **evt)
-{
-	struct bt_hci_cp_le_add_dev_to_wl *cmd = (void *)buf->data;
-	struct bt_hci_evt_cc_status *ccst;
-	u32_t status;
-
-	status = ll_filter_add(cmd->addr.type, &cmd->addr.a.val[0]);
-
-	ccst = cmd_complete(evt, sizeof(*ccst));
-	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_MEM_CAPACITY_EXCEEDED;
-}
-
-static void le_rem_dev_from_wl(struct net_buf *buf, struct net_buf **evt)
-{
-	struct bt_hci_cp_le_rem_dev_from_wl *cmd = (void *)buf->data;
-	struct bt_hci_evt_cc_status *ccst;
-	u32_t status;
-
-	status = ll_filter_remove(cmd->addr.type, &cmd->addr.a.val[0]);
-
-	ccst = cmd_complete(evt, sizeof(*ccst));
-	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED;
-}
-
-static void le_conn_update(struct net_buf *buf, struct net_buf **evt)
-{
-	struct hci_cp_le_conn_update *cmd = (void *)buf->data;
-	u16_t supervision_timeout;
-	u16_t conn_interval_max;
-	u16_t conn_latency;
-	u32_t status;
-	u16_t handle;
-
-	handle = sys_le16_to_cpu(cmd->handle);
-	conn_interval_max = sys_le16_to_cpu(cmd->conn_interval_max);
-	conn_latency = sys_le16_to_cpu(cmd->conn_latency);
-	supervision_timeout = sys_le16_to_cpu(cmd->supervision_timeout);
-
-	/** @todo if peer supports LE Conn Param Req,
-	* use Req cmd (1) instead of Initiate cmd (0).
-	*/
-	status = ll_conn_update(handle, 0, 0, conn_interval_max,
-				conn_latency, supervision_timeout);
-
-	*evt = cmd_status((!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED);
-}
-
 static void le_set_host_chan_classif(struct net_buf *buf, struct net_buf **evt)
 {
 	struct bt_hci_cp_le_set_host_chan_classif *cmd = (void *)buf->data;
@@ -696,43 +710,6 @@ static void le_set_host_chan_classif(struct net_buf *buf, struct net_buf **evt)
 
 	ccst = cmd_complete(evt, sizeof(*ccst));
 	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED;
-}
-
-static void le_read_remote_features(struct net_buf *buf, struct net_buf **evt)
-{
-	struct bt_hci_cp_le_read_remote_features *cmd = (void *)buf->data;
-	u32_t status;
-	u16_t handle;
-
-	handle = sys_le16_to_cpu(cmd->handle);
-	status = ll_feature_req_send(handle);
-
-	*evt = cmd_status((!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED);
-}
-
-static void le_encrypt(struct net_buf *buf, struct net_buf **evt)
-{
-	struct bt_hci_cp_le_encrypt *cmd = (void *)buf->data;
-	struct bt_hci_rp_le_encrypt *rp;
-	u8_t enc_data[16];
-
-	ecb_encrypt(cmd->key, cmd->plaintext, enc_data, NULL);
-
-	rp = cmd_complete(evt, sizeof(*rp));
-
-	rp->status = 0x00;
-	memcpy(rp->enc_data, enc_data, 16);
-}
-
-static void le_rand(struct net_buf *buf, struct net_buf **evt)
-{
-	struct bt_hci_rp_le_rand *rp;
-	u8_t count = sizeof(rp->rand);
-
-	rp = cmd_complete(evt, sizeof(*rp));
-	rp->status = 0x00;
-
-	bt_rand(rp->rand, count);
 }
 
 static void le_start_encryption(struct net_buf *buf, struct net_buf **evt)
@@ -781,14 +758,39 @@ static void le_ltk_req_neg_reply(struct net_buf *buf, struct net_buf **evt)
 	rp->handle = sys_le16_to_cpu(handle);
 }
 
-static void le_read_supp_states(struct net_buf *buf, struct net_buf **evt)
+static void le_read_remote_features(struct net_buf *buf, struct net_buf **evt)
 {
-	struct bt_hci_rp_le_read_supp_states *rp;
+	struct bt_hci_cp_le_read_remote_features *cmd = (void *)buf->data;
+	u32_t status;
+	u16_t handle;
 
-	rp = cmd_complete(evt, sizeof(*rp));
-	rp->status = 0x00;
+	handle = sys_le16_to_cpu(cmd->handle);
+	status = ll_feature_req_send(handle);
 
-	sys_put_le64(0x000003ffffffffff, rp->le_states);
+	*evt = cmd_status((!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED);
+}
+
+static void le_conn_update(struct net_buf *buf, struct net_buf **evt)
+{
+	struct hci_cp_le_conn_update *cmd = (void *)buf->data;
+	u16_t supervision_timeout;
+	u16_t conn_interval_max;
+	u16_t conn_latency;
+	u32_t status;
+	u16_t handle;
+
+	handle = sys_le16_to_cpu(cmd->handle);
+	conn_interval_max = sys_le16_to_cpu(cmd->conn_interval_max);
+	conn_latency = sys_le16_to_cpu(cmd->conn_latency);
+	supervision_timeout = sys_le16_to_cpu(cmd->supervision_timeout);
+
+	/** @todo if peer supports LE Conn Param Req,
+	* use Req cmd (1) instead of Initiate cmd (0).
+	*/
+	status = ll_conn_update(handle, 0, 0, conn_interval_max,
+				conn_latency, supervision_timeout);
+
+	*evt = cmd_status((!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED);
 }
 
 static void le_conn_param_req_reply(struct net_buf *buf, struct net_buf **evt)
@@ -904,6 +906,35 @@ static int controller_cmd_handle(u8_t ocf, struct net_buf *cmd,
 		le_set_random_address(cmd, evt);
 		break;
 
+	case BT_OCF(BT_HCI_OP_LE_READ_WL_SIZE):
+		le_read_wl_size(cmd, evt);
+		break;
+
+	case BT_OCF(BT_HCI_OP_LE_CLEAR_WL):
+		le_clear_wl(cmd, evt);
+		break;
+
+	case BT_OCF(BT_HCI_OP_LE_ADD_DEV_TO_WL):
+		le_add_dev_to_wl(cmd, evt);
+		break;
+
+	case BT_OCF(BT_HCI_OP_LE_REM_DEV_FROM_WL):
+		le_rem_dev_from_wl(cmd, evt);
+		break;
+
+	case BT_OCF(BT_HCI_OP_LE_ENCRYPT):
+		le_encrypt(cmd, evt);
+		break;
+
+	case BT_OCF(BT_HCI_OP_LE_RAND):
+		le_rand(cmd, evt);
+		break;
+
+	case BT_OCF(BT_HCI_OP_LE_READ_SUPP_STATES):
+		le_read_supp_states(cmd, evt);
+		break;
+
+#if defined(CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV)
 	case BT_OCF(BT_HCI_OP_LE_SET_ADV_PARAM):
 		le_set_adv_param(cmd, evt);
 		break;
@@ -923,6 +954,7 @@ static int controller_cmd_handle(u8_t ocf, struct net_buf *cmd,
 	case BT_OCF(BT_HCI_OP_LE_SET_ADV_ENABLE):
 		le_set_adv_enable(cmd, evt);
 		break;
+#endif /* CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV */
 
 	case BT_OCF(BT_HCI_OP_LE_SET_SCAN_PARAM):
 		le_set_scan_param(cmd, evt);
@@ -940,40 +972,8 @@ static int controller_cmd_handle(u8_t ocf, struct net_buf *cmd,
 		le_create_conn_cancel(cmd, evt);
 		break;
 
-	case BT_OCF(BT_HCI_OP_LE_READ_WL_SIZE):
-		le_read_wl_size(cmd, evt);
-		break;
-
-	case BT_OCF(BT_HCI_OP_LE_CLEAR_WL):
-		le_clear_wl(cmd, evt);
-		break;
-
-	case BT_OCF(BT_HCI_OP_LE_ADD_DEV_TO_WL):
-		le_add_dev_to_wl(cmd, evt);
-		break;
-
-	case BT_OCF(BT_HCI_OP_LE_REM_DEV_FROM_WL):
-		le_rem_dev_from_wl(cmd, evt);
-		break;
-
-	case BT_OCF(BT_HCI_OP_LE_CONN_UPDATE):
-		le_conn_update(cmd, evt);
-		break;
-
 	case BT_OCF(BT_HCI_OP_LE_SET_HOST_CHAN_CLASSIF):
 		le_set_host_chan_classif(cmd, evt);
-		break;
-
-	case BT_OCF(BT_HCI_OP_LE_READ_REMOTE_FEATURES):
-		le_read_remote_features(cmd, evt);
-		break;
-
-	case BT_OCF(BT_HCI_OP_LE_ENCRYPT):
-		le_encrypt(cmd, evt);
-		break;
-
-	case BT_OCF(BT_HCI_OP_LE_RAND):
-		le_rand(cmd, evt);
 		break;
 
 	case BT_OCF(BT_HCI_OP_LE_START_ENCRYPTION):
@@ -988,8 +988,12 @@ static int controller_cmd_handle(u8_t ocf, struct net_buf *cmd,
 		le_ltk_req_neg_reply(cmd, evt);
 		break;
 
-	case BT_OCF(BT_HCI_OP_LE_READ_SUPP_STATES):
-		le_read_supp_states(cmd, evt);
+	case BT_OCF(BT_HCI_OP_LE_READ_REMOTE_FEATURES):
+		le_read_remote_features(cmd, evt);
+		break;
+
+	case BT_OCF(BT_HCI_OP_LE_CONN_UPDATE):
+		le_conn_update(cmd, evt);
 		break;
 
 	case BT_OCF(BT_HCI_OP_LE_CONN_PARAM_REQ_REPLY):
