+++ subsys/bluetooth/controller/ll_sw/Makefile
@@ -2,3 +2,4 @@ ccflags-y += -I$(srctree)/subsys/bluetooth/controller/include
 ccflags-y += -I$(srctree)/subsys/bluetooth/controller
 
 obj-y += crypto.o ctrl.o ll.o
+obj-$(CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV)   += ll_adv.o
